Implements Dependency Injection for C#
This is a simple dependency injection example for C#  using the notion of a Game Manager that has a Scene Manager and Audio Manager dependency. We create Interfaces for each of these dependencies and pass in instances of this interface in the constructor of the Game Manager class, injecting the dependant objects when we construct the instance of the Game Manager class and thus implementing dependency injection.